requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
openai==1.3.7
numpy==1.24.3
faiss-cpu==1.7.4
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6 